third_party/blink/web_tests/virtual/web-components-v0-disabled/shadow-dom/v0/slots-in-v0-expected.txt

CONSOLE ERROR: line 29: Uncaught TypeError: parent.createShadowRoot is not a function
